refactor: reuse button
