+++ frontend/src/views/sql-editor/EditorCommon/index.ts
@@ -3,6 +3,7 @@ import ConnectionPathBar from "./ConnectionPathBar.vue";
 import DisconnectedIcon from "./DisconnectedIcon.vue";
 import EditorAction from "./EditorAction.vue";
 import ExecutingHintModal from "./ExecutingHintModal.vue";
+import OpenAIButton from "./OpenAIButton/OpenAIButton.vue";
 import { ResultViewV1 } from "./ResultView";
 import SaveSheetModal from "./SaveSheetModal.vue";
 import SharePopover from "./SharePopover.vue";
@@ -18,6 +19,7 @@ export {
   ResultViewV1,
   DisconnectedIcon,
   SheetConnectionIcon,
+  OpenAIButton,
 };
 
 export * from "./hover-state";
